chore: move storage matter and separate the interface and the implementation
